requirements.txt
# webapp 
flask==2.2.*

# yolov7 required
opencv-python-headless
matplotlib>=3.2.2
numpy>=1.18.5
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
protobuf<4.21.3
pandas>=1.1.4
